refactor: create repositories for every feature like ui + rename resources to resource for consistency
