+++ lib/ui/auth/view_model/auth_controller.dart
@@ -1,6 +1,6 @@
 
 import 'package:devclub_app/ui/auth/view_model/auth_loading_state.dart';
-import 'package:devclub_app/data/repositories/firebase_auth_providers.dart';
+import 'package:devclub_app/data/repositories/auth/firebase_auth_providers.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
